+++ modules/mongodb/versions.tf
@@ -1,18 +1,17 @@
 terraform {
-  required_version = ">= 1.3.2"
-
+  required_version = ">=1.2.5"
   required_providers {
     aws = {
       source  = "hashicorp/aws"
       version = ">= 5.61"
     }
-    helm = {
-      source  = "hashicorp/helm"
-      version = ">= 3.0.2"
-    }
     mongodbatlas = {
       source  = "mongodb/mongodbatlas"
-      version = "~> 1.11"
+      version = ">= 1.11.0"
+    }
+    random = {
+      source  = "hashicorp/random"
+      version = ">= 3.7.2"
     }
   }
 }